% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towcab.R
\name{analyze_all_pathway_results}
\alias{analyze_all_pathway_results}
\title{analyze_all_pathway_results}
\usage{
analyze_all_pathway_results(
  deg_lists,
  out_dir,
  method_colors,
  organism = "mmusculus",
  technical_file = NULL
)
}
\arguments{
\item{out_dir}{the output directory to write all the results into}

\item{method_colors}{a list with the method names and the colors to code them by}

\item{organism}{takes the gprofiler codes ('hsapiens, mmusculus, etc')}

\item{technical_file}{a file containing the names of pathways that should be filtered out & not \
counted against a method. These should be defined as technical in nature, \
and need to have evidence through technical replicates that these are technical. \
There is a default file distributed with the package that has a draft generated \
from technical replicates of mouse brain and heart with Chromium V3.}

\item{sig_res}{a list whose names correspond to the names of the methods, coding the towcab results in that entry}
}
\value{
writes outputs in the out_dir
}
\description{
\code{analyze_all_pathway_results} Performs the full towcab method comparison and high level overview
}
