% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towcab.R
\name{do_topology_filter}
\alias{do_topology_filter}
\title{run_towcab_analysis}
\usage{
do_topology_filter(
  local_temp_mat,
  local_meta,
  local_topology_graph,
  num_boots = 100,
  abs_z_cut = 1.5,
  clip_val = 0.1
)
}
\arguments{
\item{local_temp_mat}{input matrix for only a single cluster}

\item{local_meta}{the meta data frame for only the cells contained in the pertinent cluster. A couple important notes: needs columns called "cell_id" and "label" which are consistent cell_ids across the three inputs, matching the node names in the network, colnames in the matrix, and rownames in the meta. "Label" is the batch ID.}

\item{local_topology_graph}{the kNN graph used for clustering (may either be subset of this cluster or the full network)}

\item{num_boots}{The number of bootstrap shuffled versions of the kNN used to estimate each cell's null distribution for what that cell's (default=100)}

\item{abs_z_cut}{The cutoff of the absolute value of the Z away from the mean for a cell to be excluded as non-overlapping. For example, a cell with a cross-batch connection fraction z-score > 1.5 or < -1.5 away from the bootstrap shuffled edges from the null background (default=1.5)}

\item{clip_val}{The lower clip value to include a sample; the upper clip is calculated as 1-clip_val. For example, if a sample doesn't have at least 10% connections across batch, exclude it because it could be a spurios thing that doesn't represent broad patterns. Same at 90%. (default = 0.10)}
}
\value{
a filtered version of the input matrix and meta containing only the topologically overlapped subset
}
\description{
\code{run_towcab_analysis} Performs the full towcab analysis
}
