% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towcab.R
\name{identify_any_mixed_clusters}
\alias{identify_any_mixed_clusters}
\title{run_towcab_analysis}
\usage{
identify_any_mixed_clusters(p_table, marker_list, alpha_cutoff = 1e-06)
}
\arguments{
\item{p_table}{found in tow_cab_results$p_table.}

\item{marker_list}{a list of cell-type or state marker genes that should not be topologically overlapped.}

\item{alpha_cutoff}{the significance cutoff}
}
\value{
a list of lists of lists of results. First level is cell-type-1 (noted by marker genes), second level is cell-type-2, and third level is cluster, yielding a list genes that were markers for cell-type-1 and cell-type-2 that were both significantly differentially expressed across batches in the given cluster.
}
\description{
\code{identify_any_mixed_clusters} Performs the mixed marker gene analysis
}
