% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towcab.R
\name{plot_winClust_acrossBatch_DEG_results}
\alias{plot_winClust_acrossBatch_DEG_results}
\title{run_towcab_analysis}
\usage{
plot_winClust_acrossBatch_DEG_results(sig_res, out_dir, method_colors = NULL)
}
\arguments{
\item{sig_res}{an expression matrix with cells in columns, genes in rows}

\item{out_dir}{a vector that can be interpreted as a factor corresponding to batches (same order as columns in exprs)}

\item{method_colors}{TODO}
}
\value{
a list of all results from the towcab analysis
}
\description{
\code{run_towcab_analysis} Performs the full towcab analysis
}
\examples{
   x <- rep( c( rep(1, 50), rep(2, 50) ), 2)
   y <- c(rep(1, 100), rep(2, 100))
   cont_table <- get_cont_table(x, y)
}
