% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/towcab.R
\name{run_towcab_analysis}
\alias{run_towcab_analysis}
\title{run_towcab_analysis}
\usage{
run_towcab_analysis(
  exprs,
  batch_vect,
  clust_vect,
  alpha = 0.01,
  n_threads = 10,
  min_express_for_inclusion_percentile = 0.1,
  num_pseudo_replicates = 3,
  min_cells_per_pseudo_rep = 3,
  species = "mmusculus",
  topology_filter = TRUE,
  in_topology_graph = NULL,
  method_name = ""
)
}
\arguments{
\item{exprs}{an expression matrix with cells in columns, genes in rows}

\item{batch_vect}{a vector that can be interpreted as a factor corresponding to batches (same order as columns in exprs)}

\item{clust_vect}{an integer vector (indexed either from 0 or 1) with cluster ID numbers.}

\item{alpha}{The BH corrected significance cutoff (default=0.01)}

\item{n_threads}{number of threads to use (default=10)}

\item{min_express_for_inclusion_percentile}{the percentile cutoff for determining the downsampling threshold (default is 0.10 (meaning 90% of cells will be included))}

\item{num_pseudo_replicates}{Number of pseudoreplicates for Libra DE analysis (default=3)}

\item{min_cells_per_pseudo_rep}{Minimum number of cells to include in each pseudoreplicate (default=3)}

\item{species}{The species code from gprofiler for pathway analysis (default="mmusculus", could also be something like hsapiens, etc)}

\item{topology_filter}{Boolean of whether to perform filtering for only topologically overlapping areas.}

\item{in_topology_graph}{An igraph network - this should be the kNN or similar object that went into the clustering algorithm.}

\item{method_name}{The name of the method (empty string "" by default). Not required except for printing out some things while the function is running.}
}
\value{
a list of all results from the towcab analysis
}
\description{
\code{run_towcab_analysis} Performs the full towcab analysis
}
